+++ CHANGELOG.md
@@ -15,6 +15,7 @@ FEATURES:
 * **New Data Source:** `vcd_vapp` vApp - ([#218])
 * **New Data Source:** `vcd_nsxv_dnat` DNAT for advanced edge gateways using proxied NSX-V API - [#328]
 * **New Data Source:** `vcd_nsxv_snat` SNAT for advanced edge gateways using proxied NSX-V API - [#328]
+* **New Data Source:** `vcd_independent_disk` Independent disk - [#349]
 * **New Data Source:** `vcd_catalog_media` Media item - [#340]
 
 IMPROVEMENTS:
@@ -47,6 +48,9 @@ IMPROVEMENTS:
 * `vcd_org_vdc` Add import capability and full read support [#218]
 * Upgrade Terraform SDK dependency to 0.12.8 [#320]
 * `resource/vcd_vapp_vm` has new field `computer_name` [#334]
+* Import functions can now use custom separators instead of "." [#343]
+* `resource/vcd_independent_disk` Add computed properties (`iops`, `owner_name`, `datastore_name`, `is_attached`) and read support for all fields except the `size` [#349]
+* `resource/vcd_independent_disk` Disk state ID changed from name of disk to vCD ID [#349]
 * Import functions can now use a custom separator instead of "." [#343]
 * `resource/vcd_catalog_media` Add computed properties (`is_iso`, `owner_name`, `is_published`, `creation_date`, `size`, `status`, `storage_profile_name`) and full read support [#340]
 * `resource/vcd_catalog_media` MediaItem state ID changed from colon separated list of catalog name and media name to vCD ID [#340]